backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.1
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
razorpay==1.4.1 